Field Lifecycle part 2

+++ app/Filament/Resources/CustomerResource.php
@@ -40,7 +40,7 @@ class CustomerResource extends Resource
                     ->label('Alamat')
                     ->required(),
                 TextInput::make('telepon_customer')
-                    ->integer()
+                    ->numeric()
                     ->label('Telepon')
                     ->required(),
             ]);
